postfixes
git-svn-id: https://ggnfs.svn.sourceforge.net/svnroot/ggnfs/trunk@41 80ef69ca-9341-0410-9d60-8acc7774fa31

+++ src/lasieve4/fbgen.c
@@ -31,7 +31,6 @@
 #include "ppc32/siever-config.h"
 #else
 #include "asm/lasieve-asm.h"
-
 volatile u32_t modulo32;
 #endif
 #include "lasieve.h"
@@ -240,9 +239,9 @@ void polprintmod32(u32_t * P, u32_t dP, char *c)
 {
   u32_t i;
 
-  printf("\n%s %u\n", c, dP);
+  printf("\n%s %u\n", c, (uint)dP);
   for (i = 0; i <= dP; i++)
-    printf("%u ", P[i]);
+    printf("%u ", (uint)P[i]);
   printf("\n");
 }
 
@@ -307,7 +306,7 @@ u32_t polrootmod32(mpz_t * A, u32_t dT, u32_t * r)
     r[i] = 0;
     j = polrootrecmod32(Q, dQ, r + i, S + dS + 1);      /*j=dQ */
     if (j != dQ)
-      fprintf(stderr, "Falsche Nullstellenanzahl Q, P %u\n", modulo32);
+      fprintf(stderr, "Falsche Nullstellenanzahl Q, P %u\n", (uint)modulo32);
     i += j;
   }
   if (dS > 0) {
@@ -315,7 +314,7 @@ u32_t polrootmod32(mpz_t * A, u32_t dT, u32_t * r)
     r[i] = 0;
     j = polrootrecmod32(S, dS, r + i, S + dS + 1);      /*j=dS */
     if (j != dS)
-      fprintf(stderr, "Falsche Nullstellenanzahl S, P %u\n", modulo32);
+      fprintf(stderr, "Falsche Nullstellenanzahl S, P %u\n", (uint)modulo32);
     i += j;
   }
   qsort(r, i, sizeof(u32_t), uintcmp);
